requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
pymysql
pydantic
python-dotenv
scikit-learn
tensorflow
imblearn
pandas